+++ src/app/app.module.ts
@@ -4,6 +4,7 @@ import { LeafletModule } from '@asymmetrik/ngx-leaflet';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { RisingMapModule } from './rising-map/rising-map.module';
+import { HttperrModule } from './httperr/httperr.module';
 
 @NgModule({
 	declarations: [
@@ -13,7 +14,8 @@ import { RisingMapModule } from './rising-map/rising-map.module';
 		BrowserModule,
 		LeafletModule.forRoot(),
 		AppRoutingModule,
-		RisingMapModule
+		RisingMapModule,
+		HttperrModule
 	],
 	providers: [],
 	bootstrap: [AppComponent]