+++ src/app/rising-map/rising-map.module.ts
@@ -5,12 +5,12 @@ import { LeafletModule } from '@asymmetrik/ngx-leaflet';
 import { RisingMapRoutingModule } from './rising-map-routing.module';
 
 @NgModule({
-  declarations: [RisingMapComponent],
-  imports: [
-    CommonModule,
-    LeafletModule,
-    RisingMapRoutingModule
-  ],
-  exports: [RisingMapComponent]
+	declarations: [RisingMapComponent],
+	imports: [
+		CommonModule,
+		LeafletModule,
+		RisingMapRoutingModule
+	],
+	exports: [RisingMapComponent]
 })
 export class RisingMapModule { }