+++ src/environments/environment.omega.ts
@@ -1,0 +1,5 @@
+export const environment = {
+	production: true,
+	mapTileUrl: "https://{s}.omega-zirkel.de/{z}/{x}/{y}.png",
+	mapTileSubdomains: ['tiles01', 'tiles02', 'tiles03', 'tiles04', 'tiles05', 'tiles06']
+};