+++ HISTORY.md
@@ -1,8 +1,18 @@
 ## [Unreleased]
 
+## [0.3.0] - 2019-05-12
+### Added
+- environment configuration setting `omega`
+- `mapTileSubdomains` array in all environment files
+- refresh button and refresh interval  control in the bottom left corner
+- cacheBuster for image files (refresh does not work without a buster)
+- 2 new routes `:posX/:posY` and `:posX/:posY/zoom` to bookmark and share map locations
+- added component for 404 Errors
+
 ## [0.2.0] - 2019-03-06
 ### Added
 - Ingame coordinates are now displayed in the bottom left corner
+- 2 more zoomlevel without new tiles (zoom into origin tiles)
 
 ## [0.1.0] - 2019-03-02
 ### Initial